requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.17.3
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
